application controller renamed and parent changed
The gem was obliterating the rails app app_controller.
