+++ lib/supplejack_api/engine.rb
@@ -35,7 +35,7 @@ module SupplejackApi
     end
 
     config.to_prepare do
-      ApplicationController.helper(ActionView::Helpers::ApplicationHelper)
+      SupplejackApplicationController.helper(ActionView::Helpers)
       Dir.glob(Rails.root + 'app/decorators/**/*_decorator*.rb').each do |c|
         require_dependency(c)
       end
